pyproject.toml
[tool.poetry]
name = "StockStream Insight"
version = "0.1.0"
description ="A cutting-edge data pipeline designed with medallion architecture to demonstrate real-time stock price analysis."
authors = ["Krit Pragobdee <krit.pragobdee@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
finnhub-python = "^2.4.19"
pandas = "^2.2.0"
torch = "^2.2.0"
pyarrow = "^15.0.0"
pyspark = "^3.5.0"
transformers = "^4.37.2"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
azure-eventhub = "^5.11.6"
pytz = "^2024.1"
setuptools = "^69.1.0"
wheel = "^0.42.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
ruff = "^0.2.2"
pytest-cov ="^4.1.0"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".databricks",
    ".githubs",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
